+++ dojo/tools/clair/clairklar_parser.py
@@ -1,5 +1,5 @@
 
 
-class ClairKlarParser(object):
+class ClairKlarScan(object):
     def __init__(self) -> None:
         pass
